Move Phrase and impls to separate file
